In all the apache-php application, in _prep_<package> function, cleaning of staging directory was not proper as, it contains staging/<platform>              -- <package>              -- installer/              -- scripts/ however, it is removing/cleaning only staging/<platform>/<package>.
The error came to picture when building pgphonehome on the new buildfarm as pgphonehome as of now doesn't require scripts directory but it is being included in the installer.xml.in. It dint threw any error in the current buildfarm because of improper cleaning of staging directory.

Commited changes for above.