Fix what we believe is one of the more common outstanding failure cases for installation.
When we run the installer the working directory for the various installation scripts is
typically the directory containing the installer that the user just double-clicked. This
isn't a problem for any of them, except for initcluster.vbs when it runs initdb. Because
we never run posgres.exe with administrative permissions, initb.exe re-exec's itself under
a restricted security token before it does anything else. If the working directory for the
script is a secure location, we may no longer be able to access it with the restricted
token, so postgres.exe exits with an error because the current directory isn't valid. This
causes initdb to fail entirely.

The fix is to have initcluster.vbs change the working directory to the installation
directory - if we cannot work there, things are going to go badly wrong anyway. We
also cannot assume we can access %TEMP%, so initdb's temporary password file needs to go
in the installation directory as well.

Thanks to Magnus Hagander for a branstorming session, and testing by Daniel <who works
in a secure lab we cannot mention!>