fixes #36432, refs #32485
As part of fix for 32485, the Base Directory path in the registry contains the forward slashes
instead of backward slashes because of the absolutize action of bitrock installbuilder. This
makes the installation directory validation check to fail.
Revert that fix and add the fix to correct the slashes in the Base Directory key value.