[server]
Set server version to 12.0beta3, pgAdmin version to 4.11.

This is the first server version to be built using OpenSSL 1.1.1c. Make necessary changes
in the server build scripts to handle the change in the new SSL library names. Also, make changes in the
installer to remove the older SSL libraries in upgrade mode installation

Bundle latest VC++ 2017 vcredist on Windows

Fix the security issue on Windows. The vulnerability was reported in the windows installer.
To create the cluster the installer was creating a file to store the password in
the directory that was workd readable. The changes are made in the installer so that the password file is created in the directory with limited ACLs.

Reviewed by Sandeep Thakkar