fixes #DB-326 and other related issues. Bump server build number to generate new builds.
Actually, we made a change in the code-signing step (we added entitlement) for our binaries
and regenerated the installer. On fresh installation, the binaries work fine. On an upgrade
to the existing installation of previous version, we observed the same binaries do not work
but they do work fine when we reboot the server.

$ /Library/PostgreSQL/12/bin/pg_ctl
Killed: 9
$ sudo reboot
$ /Library/PostgreSQL/12/bin/pg_ctl --version
pg_ctl (PostgreSQL) 12.2

This was posted on Apple developer forums and got to know that it happens because kernel caches
information about the signature and doesn’t flush that cache when you rewrite the Mach-O image file.
The solution is to not rewrite the image file, but instead create a new image file and move it into place.

This was reported to Bitrock to confirm how the overwrite of files happen.
Bitrock has provided us with the fix in their InstallBuilder tool. So, no change in the installer
scripts for the -3 builds