remove prerun_checks.vbs as it is not required
This script checks if the %TEMP% is writeable and
if the script can be run from inside the TEMP location.
Since the use of TEMP location is evolved over years and
we have createTempDir function that creates a sub-directory
inside TEMP rather than writing in the TEMP directory directly,
so we do not need this script now.

DBP-510, Manika Singhal --Tested by Manika Singhal --Reviewed by Sandeep Thakkar