Discontinue to bundle debug-symbols in PG installers and their binaries.
Historically, we are bundling *.pdb files in our PG installers binaries zip and exe files.
But now we are reconsidering it, given there size is increasing
like in PG15 size was ~110MB and on PG17 size is ~158MB.
So, we are discontinuing to bundle them with installer and binaries.
We maintain a separate archive of .pdb files and keep that on CS instead,
so if any customer demands for it we can always share it with them.
In EPAS, we are following this later way.

--Manika Singhal --Reviewed by --Sandeep Thakkar, Srinu Perabattula