Re-write initcluster.vbs into powershell script and fix the initdb failure
when the locale passed through the installer script contains non-ascii characters

The commit handles two issues in a single commit:

With Microsoft announcing the retirement of VBScript, this commit replaces the
last used VB script (initcluster.vbs) to a powershell script (initcluster.ps1).

It also addresses the issue of database initialisation failure if locale contains
non-ASCII characters. During installation, the EDB GUI installer would run
getlocales.exe retrieving a list of all the available system locales, populate the
list in a drop down box for users to choose the desired locale. The locale
containing the non-ascii characters when passed to initdb.exe returns the
error. This behaviour changes with this commit

Replace getlocales.exe with a powershell command to return the list of
locales. When user chooses a locale, it'll now be converted into BCP-47 code
name before passing to initcluster script and further to initdb.exe.

JIRA Issues: DBP-719, DBP-788,
edb-installers/issues - #127, #148, #186, #193, #220
--Manika Singhal, Reviewed by- Dave Page, Sandeep Thakkar, 
    Tested By- Manika Singhal, Kashif Zeeshan