set PG_SYSROOT instead of -isysroot
When -isysroot is added to CFLAGS, anyone who tries to compile
extensions against the server headers will have to use the same SDK
version and location as the server had, which may not be possible.

If the SDK path is placed into PG_SYSROOT instead, anyone who builds an
extension against the server may point it to a setting that is more
fitting to their environment.

This has been reported a few times by users of pgvector, with one of the
latest being https://github.com/pgvector/pgvector/issues/592#issuecomment-2666735840