Improvements in DoCmd function in initcluster script
No more temporary files: The variables $scriptFile, $outputFile,
and the associated Set-Content and Remove-Item calls are all gone.
The function no longer needs to write and execute an external script.

Direct Execution: The command is executed directly using cmd.exe /c.
to ensure the output streams and exit code are correctly captured.
The captured output is stored in a variable and then displayed to the
console, providing a clear view of what the command did. (Issue 347)

Also, disable wrapInScript for getlocales.ps1 as it was failing in
few conditions (Issue 343)

Issue #347 Issue #343