Update initcluster.ps1 to convert "English, <Country> with English_<Country>"
The getlocales script returns the long names in the format "English, <Country>"
but it seems the Windows C library does not support it and requires it in 
English_<Country>.<CodePage> format. Hence, when any English long name 
locale is passed to initdb.exe in this format, it automatically sets the locale to 
"English_United States.1252"

This is reported by Ben Casp in pgsql-bugs and theory supported by Thomas Munro at: https://www.postgresql.org/messageid/CANFyU9642ZG81fmUPp7ceUNP%2Bb7ZnpWzAuzsTqoy5p1H%2BQEpag%40mail.gmail.com

As a quick work-around/fix, make changes in the initcluster script to convert 
the "English, <Country> with English_<Country>" before passing to initdb.exe

It was also found the installer is not populating the locales list when the getlocales
script is not wrapped i.e when wrapInScript is set to 0. Hence, set to 1 again and 
that reverts the fix done for #Issue 343. This will be addressed separately later

Sandeep Thakkar, Reviewed by Kritika Agarwal

DBP-1529